requirements.txt
nicegui
pyserial
crcmod
